Models imported, Player movment, tile generation, and framework for hazard and collision handling
Models for the player, some tile decorations, and vehicles were added.

Player currently exists as a capsule and the movement script contains almost all of the movement handling code, but does not yet have inputs applied.

Tiles now do have a manager that generates both at the start and on demand, but requires further testing. currently there is no way to trigger generation of the next row of tiles.

obstacle framework is designed but empty. obstacles will use a public enum to designate type. if i had more time id love to make this into a scriptable object, but this will do for now.